Created Basic Player Controller that Interacts with Interactable
Created basic player controller that changes the cursor and if it finds any interactables in front of the player it will interact with the object. Created a (temporary) Interactable class and created Water Manager which inherits from the interactable class. I created a (temporary) state machine to manage the three different states of water (its a complex solution to a simple problem but I wanted to show off a possibility of creating a water state manager).