Added Basic Functional Water States
Altered Interactable to IInteractable because its a interface now instead of an abstract class. I created a Object Pooler to instantiate and manage the steam and water objects in the level. In Addition I created a game manager singleton to act as an accessor so that objects can get access to the Object Pooler if they need to.