Updated Object Pooler and Spawners
I made it so pooled object's scale can be specified. I made it so that the designer can either use the spawner's transform or they can use the script component's variables instead. I also made the spawner's better display what you are spawning and for the GUIs to be shown constantly instead of only when selected.